refactor(smtp_client,imap_server): splitter en sous-modules (≤400 LOC)
- smtp_client/mod.rs 538→78 : extract relay.rs, mx.rs, session.rs
- imap_server/mod.rs 529→111 : extract commands.rs, parser.rs
